add registration and likes

+++ src/components/DeleteArticle.jsx
@@ -7,22 +7,26 @@ import { deleteObject, ref } from "firebase/storage";
 export default function DeleteArticle({id, imageUrl}) {
 
     const handleDelete=async()=>{
-        try {
-           await  deleteDoc(doc(db, "Articles", id))
-           toast("Article deleted successfully", { type: "success" });
-           const storageRef = ref(storage, imageUrl);
-            await deleteObject(storageRef);
-        } catch (error) {
-            toast("Error deleting article", { type: "error" });
-            console.log(error);
-        }
-       
-    }
+        if (window.confirm("Are you sure you want to delete this article?")) {
+            try {
+              await deleteDoc(doc(db, "Articles", id));
+              toast("Article deleted successfully", { type: "success" });
+              const storageRef = ref(storage, imageUrl);
+              await deleteObject(storageRef);
+            } catch (error) {
+              toast("Error deleting article", { type: "error" });
+              console.log(error);
+            }
+          }
+        };
 
   return (
     <div>
-        <button className='btn btn-danger'
-        onClick={handleDelete}>Delete</button>
+       <i
+        className="fa fa-times"
+        onClick={handleDelete}
+        style={{ cursor: "pointer" }}
+      />
     </div>
   )
 }
